The script is not developed by me, but after struggling finding elegant  solution I found this in deeps of googles results.
The last message in the forum was 2005, so I decided rather keep this in live.

This Apple script can be easily copied into Automator, saved as folder service, install into ~/Library/Services and we can import on right click on file or folder.
(prerequisite is install flac until - I found and installed from macports - just changed the paths at line 29 and 30) that's it.

Many thanks to original author.